Add alarm table delegate, make alarms persistent. After this commit I'm going to redo everything I did and make the picker on the same page.
Explanation: So you know how every time you switch to a new view, you create a new instance of the FlowManager? You know how I made the alarms persistent to get around this? I didn't realize that I would also have to make the alarmTable persistent as well. I don't want to deal with that, so I'm going to make the picker in the same page.